update groups screen to use functional component
